src/result/CVE-2021-21330/new-python-backend/requirements.txt
aiohttp==3.9.1
aiohttp-apispec==2.2.3
aiohttp-autoreload==0.0.1
aiohttp-middlewares==2.2.1
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
apispec==3.3.2
astroid==3.0.2
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.29.0
attrs==23.1.0
black==23.12.1
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
dill==0.3.7
dnspython==2.4.2
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
isort==5.13.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pycparser==2.21
pylint==3.0.3
python-dotenv==1.0.0
PyYAML==6.0.1
SQLAlchemy==2.0.24
tomlkit==0.12.3
typing_extensions==4.9.0
ujson==5.9.0
webargs==5.5.3
yarl==1.9.4
